Move test types to testDomain file, avoid SB error.
